Improve CodeCoverage to feature/folly_futures (#34)

+++ src/include/nuraft_mesg/nuraft_mesg.hpp
@@ -17,7 +17,6 @@
 #include <list>
 #include <memory>
 #include <string>
-#include <system_error>
 #include <vector>
 
 #include <libnuraft/nuraft.hxx>
@@ -57,7 +56,7 @@ class Manager {
 public:
     struct Params {
         boost::uuids::uuid server_uuid_;
-        uint32_t mesg_port_;
+        uint16_t mesg_port_;
         group_type_t default_group_type_;
         std::string ssl_key_;
         std::string ssl_cert_;
